code-indexing/pyproject.toml
[project]
name = "code-embedding"
version = "0.1.0"
description = "Simple example for cocoindex: build embedding index based on source code."
requires-python = ">=3.10"
dependencies = ["cocoindex>=0.1.42", "python-dotenv>=1.0.1", "psycopg_pool", "psycopg", "numpy<2"]

[tool.setuptools]
packages = []